==> 03_patchformation.Rmd <==

# Patch Formation

## Set Up

### Libraries
```{r libraries2, message=FALSE, warning=FALSE}
library(tidyverse)
library(terra)
library(patchwoRk)
library(sf)
library(mapview)
library(exactextractr)
library(lubridate)
```

### Import RdNBR Rasters

```{r raster import}
# import calculated RdNBR rasters for each fire boundary polygon
rast_list <- list.files(path = "data/rdnbr_rasters", pattern='.tif', all.files=TRUE, full.names=TRUE)
rast_all <- lapply(rast_list, rast)
rast_collection <- sprc(rast_all)

crs <- crs(rast_collection[1])
```

### Import Fire Boundaries

```{r fire import, message=FALSE, warning=FALSE, results='hide'}
# import fire boundaries
mtbs_export <- st_read('data/fire_boundaries/mtbs_export.shp') %>% 
  st_transform(., crs=crs) 

fires_export <- st_read("data/fire_boundaries/fires_export.shp")%>% 
  st_transform(., crs=crs)

# import forest type group raster
conus_forestgroup <- raster('data/forest_type/conus_forestgroup.tif')
forest_codes <- read_csv('data/forest_type/forestgroupcodes.csv')
```

## Delineate High-Severity Patches

### PatchMorph
```{r patchmorph, message=FALSE, warning=FALSE, results='hide', cache=TRUE}
# loop through RdNBR rasters, assign >640 to high severity category
# utilize patchmorph to act as 3x3 cell majority filter
patch_df <- list()
for (i in 1:length(rast_all)){
  rast_fire <- raster(rast_collection[i])
  rast_fire[rast_fire < 640] <- 0
  rast_fire[rast_fire >= 640] <- 1

  patch <- patchMorph(rast_fire, spurThresh = 3, gapThresh = 3)
  patch_poly <- as.polygons(rast(patch)) %>%
    st_as_sf()
  df_union_cast <- patch_poly %>%
    st_cast(., "POLYGON") %>%
    filter(layer == 1)
  patch_df[[i]] <- df_union_cast}

patch_poly_all <- do.call(rbind,patch_df)
```

### Filter Patch Dataset

```{r patch filter, message=FALSE, warning=FALSE}
# filter small patches
patches_full <- patch_poly_all %>% 
  mutate(patch_area_ha = as.numeric(st_area(.))/10000) %>%
  filter(patch_area_ha > 2.25)
```

```{r inform patches, message=FALSE, warning=FALSE, results='hide', cache=TRUE}
# join patches back to grouped fires
patches_joined <- st_join(patches_full,mtbs_export,join = st_intersects,left= FALSE,largest = TRUE) %>%
  dplyr::select(-layer,-BurnBndAc) %>% 
  left_join(.,exact_extract(conus_forestgroup,., 'mode', append_cols = TRUE, max_cells_in_memory = 3e+08))%>%
  mutate(patch_foresttype = case_when(mode==200 ~ "Douglas-Fir",
                                     mode==220 ~ "Ponderosa",
                                     mode==260 ~ "Fir-Spruce",
                                     mode==280 ~ "Lodegepole Pine",
                                     mode==0 ~ "Unforested",
                                     TRUE ~ "Other"))
```

### Export Data 

```{r export patches, message=FALSE, warning=FALSE}
patches <- patches_joined %>%
  st_transform(crs = crs)

# st_write(patches, "data/patches/", "highsev_patches.shp",driver = 'ESRI Shapefile')
```

## Mapping

```{r map high severity patches}
mapview(patches_joined,col.regions = "red",layer.name = "High-severity Burn Patches") + 
  mapview(fires_export, alpha.regions = 0, lwd = 2,color = "black",col.regions = "black", layer.name = "Fire Boundaries")
```
