==> 07_model.Rmd <==
# Random Forest Model Development

## Set Up

### Libraries

```{r libraries5, message=FALSE, warning=FALSE}
library(mapview)
library(sf)
library(terra)
library(tidyverse)
library(ggplot2)
library(car)
library(forcats)
library(randomForest)
library(raster)
```

### Import Fire Boundaries

```{r fire boundary import, message=FALSE, warning=FALSE,results='hide'}
fires_export <- st_read("data/fire_boundaries/fires_export.shp")%>% 
  st_transform(., crs="EPSG: 4326")

crs <- crs(fires_export)
```

### Import Training Points

```{r import points, message=FALSE, warning=FALSE,results='hide'}
# bring in training points
points <- st_read("data/points/points_export.shp") %>% 
  st_transform(crs=crs)%>% 
  drop_na()
```

## Prepare Imagery

### Prepare Data for Bands and Fire Names

#### Bands
```{r band names1, message=FALSE, warning=FALSE}
# list band names and years
bands <- c("blue", "green", "red", "nir", "swir1", "swir2", "ndvi", "ndwi", "nbr", "nbr2", "ndsi", "ndfsi", "evi")
years <- c(1984:2020)

# create list of all combinations of bands, in the appropriate order
bandlist <- outer(bands, years, paste, sep="")
```

#### Fire Names

```{r fire names2, message=FALSE, warning=FALSE}
# get list of fire names
fire_names_all <- unique(fires_export$Fire_ID)
fire_names_points <- unique(points$Fire_ID,na.rm=TRUE)
```

### Merge Large Rasters

```{r merge function, message=FALSE, warning=FALSE,results='hide'}
# function to merge rasters, if there are multiple in the folder
export_rasters <- function(fire_name){
  print(paste0("Starting Fire ",fire_name))
  
  # get list of this fire's tif files
  rast_list <- list.files(path = "data/landsat/landsat_annual", pattern=fire_name, all.files=TRUE, full.names=TRUE)
  
  # larger rasters were exported from GEE as multiple files, need to be combined before importing
  if (length(rast_list)>1) { 
    print(paste0(length(rast_list)," rasters, merging..."))
    rast_all <- lapply(rast_list, rast)
    rast_collection <- do.call(merge,rast_all)
    writeRaster(rast_collection, str_c("data/landsat/landsat_annual/",fire_name,".tif"), overwrite=FALSE,gdal="COMPRESS=NONE")
    } else { 
    print("Only one raster, can extract directly")
    }
}
```

```{r message=FALSE, warning=FALSE,eval=FALSE}
# map across all fire events, merging rasters where there are multiple
map(fire_names_all,export_rasters)
```

## Extract Landsat Data 

```{r extract function, message=FALSE,warning=FALSE}
# function to extract the snow-cover landsat values for the training data points
extract_landsat <- function(fire_name){
  
  print(paste0("Starting Fire ",fire_name))
  
  # get this fire's tif files
  rast_list <- list.files(path = "data/landsat/landsat_training", pattern=str_c(fire_name,".tif"), all.files=TRUE, full.names=TRUE)
  rast_fire <- rast(rast_list)  

  # name the bands
  names(rast_fire) <- bands
  
  # verify crs
  crs(rast_fire) <- "EPSG: 4326"
  
  # filter the points for this fire
  fire_points <- points %>% 
    filter(Fire_ID==fire_name) %>% 
    st_transform(crs=crs(rast_fire))
  
  # get the mean landsat values for each patch in this fire
  extracted_points <- st_as_sf(terra::extract(rast_fire, fire_points,bind = TRUE))
  
  # export
  return(extracted_points)
}
```

```{r extract, message=FALSE, warning=FALSE,results = 'hide'}
# extract landsat values to each training point
extracted_df <- map(fire_names_points, extract_landsat)
```

### Prepare Dataset

```{r clean dataset}
# compile dataset and clean dataset
training_dataset <- do.call(rbind,extracted_df) %>% 
  mutate(absence = as.factor(case_when(class == "absence" ~ "absence",
                                       TRUE ~ "presence")),
         binom = as.factor(case_when(class == "absence" ~ 0,
                                       TRUE ~ 1)),
         class = case_when(class == "presence20to40" ~ "20-60%",
                           class == "presence40to60" ~ "20-60%",
                           class == "presence10to20" ~ "10-20%",
                           class == "presence1to10" ~ "1-10%",
                           class == "presencetrace" ~ "<1%",
                           class == "presence60plus" ~ ">60%",
                           TRUE ~ "absence"),
         class = fct_relevel(as.factor(class),c("absence","<1%","1-10%","10-20%","20-60%",">60%"))) %>% 
  st_drop_geometry() %>% 
  dplyr::select(-qd_vnt_,-qd_fr_d) %>% 
  drop_na(ndvi)
```

## Examine Data 

### Plot NDVI by Density Class

```{r ggplotting, message=FALSE, warning=FALSE}
# plot the NDVI of the training data points by visually estimated conifer cover
ggplot(training_dataset,aes(class,ndvi)) +
  geom_boxplot() + 
  labs(title = "Training Data NDVI by Percent Conifer Cover Class",x = "Visually Estimated Conifer Percent Cover",y = "NDVI") +
  theme_classic()

ggplot(training_dataset %>% filter(class %in% c("absence","<1%")),aes(class,ndvi)) +
  geom_boxplot()+
  ylim(-.1,.1) + 
  labs(title = "Training Data NDVI by Percent Conifer Cover Class", subtitle = "Comparing Trace Conifer Cover Class vs. Absence",x = "Visually Estimated Conifer Percent Cover",y = "NDVI") +
  theme_classic()

ggplot(training_dataset,aes(absence,ndvi)) +
  geom_boxplot() + 
  labs(title = "Training Data NDVI by Percent Conifer Cover Class ", subtitle = "Comparing Conifer Presence vs. Absence",x = "Conifer Status",y = "NDVI")+
  theme_classic()
```

## Develop Model

### Create Random Forest Model

```{r rf model}
# create RF model of conifer presence-absence based on snow-cover landsat image bands
rf_conifer <- randomForest(binom ~ red + green + blue + nir + swir1 + swir2 + ndsi + ndfsi + ndvi + evi + nbr + nbr2 + ndwi, data = training_dataset %>% drop_na())
```

### Evaluate Parameters

```{r rf model importance}
# model object
rf_conifer

# variable importance
randomForest::importance(rf_conifer)

# confusion matrix
rf_conifer$confusion

# summary
summary(rf_conifer)
```
### Export Model Object
```{r rf model export, eval = FALSE}
# export 
# saveRDS(rf_conifer,"data/models/rf_conifer.rds")
```
