==> 10_lei.Rmd <==

```{r}
library(sf)
library(terra)
library(mapview)
library(raster)
library(tidyverse)
library(ggplot2) 

```

```{r,eval=FALSE}
# # rename files
# rast_list <- list.files(path = "data/prediction_rasters", pattern = "rf", all.files=TRUE, full.names=TRUE)
# 
# file.rename(rast_list,sub("year1to3*", "t0", rast_list))
# file.rename(rast_list,sub("year4to6*", "t1", rast_list))
# file.rename(rast_list,sub("year7to9*", "t2", rast_list))
# file.rename(rast_list,sub("year10to12*", "t3", rast_list))
# file.rename(rast_list,sub("year13to15*", "t4", rast_list))
# file.rename(rast_list,sub("year16to18*", "t5", rast_list))
# file.rename(rast_list,sub("year19to21*", "t6", rast_list))
# file.rename(rast_list,sub("year22to24*", "t7", rast_list))
# file.rename(rast_list,sub("year25to27*", "t8", rast_list))
# file.rename(rast_list,sub("year28to30*", "t9", rast_list))
```


```{r}
fire_list <- c("Fire_1_1988","Fire_2_1988","Fire_3_1988","Fire_4_1988","Fire_7_1988","Fire_9_1988","Fire_10_1988","Fire_11_1988","Fire_12_1988","Fire_13_1988","Fire_14_1988","Fire_15_1988","Fire_16_1988","Fire_18_1988","Fire_19_1988","Fire_20_1988","Fire_22_1988","Fire_23_1988","Fire_25_1988","Fire_26_1988","Fire_28_1988","Fire_29_1988","Fire_31_1988","Fire_32_1989","Fire_33_1989","Fire_35_1989","Fire_38_1989","Fire_41_1989","Fire_42_1989","Fire_48_1990","Fire_49_1991","Fire_50_1991","Fire_51_1991","Fire_54_1991")

patches <- st_read("data/patches/highsev_patches.shp") %>% 
  mutate(Patch_ID = str_c(Fire_ID,"-",1:n()))
```


```{r message=FALSE, warning=FALSE}
get_lei <- function(fire_name){
  print(fire_name)
  rast_list <- list.files(path = "data/prediction_rasters", pattern = str_c(fire_name,"_rf"), all.files=TRUE, full.names=TRUE)
  rast_names <- str_sub(rast_list,start = -6, end = -5)  
  rast_fire <- rast(rast_list)
  names(rast_fire) <- rast_names

  patch_edge <- patches %>% 
    filter(Fire_ID == fire_name) %>% 
    st_cast(.,"LINESTRING") %>% 
    st_transform(crs = "EPSG: 3857")

  fire_df <- list()
  
  loop_names <- rast_names[1:9]
  
  for(i in loop_names){
    # second time point 
    print(i)
    i2 <- str_c("t",as.numeric(str_sub(i,2,2))+1)
      
    # set up rasters
    step1 <- rast_fire[[i]]  %>% 
      as.factor() %>% 
      terra::project(.,"EPSG: 3857")
    step2 <- rast_fire[[i2]] %>% 
      as.factor() %>% 
      terra::project(.,"EPSG: 3857")
    
        # set area outside of patches to "forested"
    ## later change this to ndvi that is reliably tree covered in snow
    step1[is.na(step1)] <- 2
    
    # convert to polygons
    step1_polygons <- as.polygons(step1) %>%
        st_as_sf() %>%
        st_cast("POLYGON") %>% 
        filter(.[[1]] == 2)
  
    step2_polygons <- as.polygons(step2) %>%
        st_as_sf() %>%
        st_cast("POLYGON") %>% 
        filter(.[[1]] == 2) 
    
    lei_df <- list()
    
    if(nrow(step2_polygons)>0){
      for(j in 1:nrow(step2_polygons)){
        print(str_c(fire_name," : ",i," : ",100*round(j/nrow(step2_polygons),2),"%"))
        step2_poly <- step2_polygons[j,]
      
        step2_newgrowth <- st_difference(step2_poly,st_union(st_as_sf(step1_polygons)) %>%  st_make_valid())
        
        step2_newgrowth_buffer <- st_as_sf(buffer(vect(step2_newgrowth),150))
        
        step2_newgrowth_buffer_diff <- st_difference(st_union(step2_newgrowth_buffer) %>% st_make_valid(),
                                                     st_union(step2_newgrowth) %>%  st_make_valid())
        
        
        if(nrow(step2_newgrowth)>0){
          lei <- as.data.frame(exactextractr::exact_extract(step1,step2_newgrowth_buffer_diff)) %>% 
            group_by(value) %>% 
            summarize(n=n()) %>% 
            ungroup() %>% 
            mutate(lei = n/sum(n),
                   area_ha = as.numeric(st_area(step2_newgrowth))/10000,
                   edge_dist_m = as.numeric(min(st_distance(st_centroid(step2_newgrowth),patch_edge))),
                   timepoint = i) %>% 
            filter(value == 2)
        } else {
          lei <- data.frame(row.names = c("value","n","lei","area-ha","edge_dist_m","category"))
        }
        lei_df[[j]] <- lei
      }
      }
      else {
        print("no new growth")
      }
  
    lei_data <- do.call(rbind,lei_df) %>% 
      mutate(category = case_when(lei > 0.5 ~ "infill",
                                  lei < .01 ~ "leapfrog",
                                  TRUE ~ "expansion"),
             area_ha = round(area_ha,2),
             edge_dist_m = round(edge_dist_m,2),
             Fire_ID = fire_name) 
    fire_df[[i]] <- lei_data
  }
return(fire_df)
}
```

```{r message=FALSE, warning=FALSE}
# test <- get_lei("Fire_48_1990")
# test2 <- get_lei("Fire_32_1989")
# test3 <- get_lei("Fire_26_1988")
# test4 <- get_lei("Fire_10_1988")
# 
# df <- do.call(rbind,test3)
# 
# bigdata <- rbind(do.call(rbind,test),do.call(rbind,test2),do.call(rbind,test3))

lei_dataset <- map(fire_list,get_lei)
```


```{r message=FALSE, warning=FALSE}
bigdata_dist <- bigdata %>% 
  group_by(timepoint,Fire_ID) %>% 
  summarize(edge_dist_m = mean(edge_dist_m))

ggplot(bigdata_dist,aes(timepoint,edge_dist_m,group= Fire_ID,color = Fire_ID)) + 
  geom_point() + 
  geom_smooth()

ggplot(bigdata %>% group_by (timepoint) %>%  summarize(edge_dist_m = mean(edge_dist_m)),aes(timepoint,edge_dist_m)) + geom_point()

ggplot(bigdata %>% group_by (timepoint) %>%  summarize(edge_dist_m = mean(edge_dist_m)),aes(timepoint,edge_dist_m)) + geom_point()

```
