==> 11_patchfate_0to30.Rmd <==

# Landscape-Level Conifer Recolonization

## Set Up

### Libraries

```{r libraries9, message=FALSE, warning=FALSE}
library(tidyverse)
library(terra)
library(sf)
library(ggsankey)
library(gridExtra)
library(rempsyc)

sf_use_s2(FALSE)
```

### Import High-Severity Patches

```{r import patch3, message=FALSE, warning=FALSE,results='hide'}
# fire list
fire_list <- c("Fire_7_1988","Fire_9_1988","Fire_10_1988","Fire_11_1988","Fire_12_1988","Fire_13_1988","Fire_14_1988","Fire_15_1988","Fire_16_1988","Fire_18_1988","Fire_19_1988","Fire_20_1988","Fire_22_1988","Fire_23_1988","Fire_25_1988","Fire_26_1988","Fire_28_1988","Fire_29_1988","Fire_31_1988","Fire_32_1989","Fire_33_1989","Fire_35_1989","Fire_38_1989","Fire_41_1989","Fire_42_1989","Fire_48_1990","Fire_49_1991","Fire_50_1991","Fire_51_1991","Fire_54_1991","Fire_1_1988","Fire_2_1988","Fire_3_1988","Fire_4_1988")

# import high-severity patches
patches <- st_read("data/patches/highsev_patches.shp") %>% 
  mutate(Patch_ID = str_c(Fire_ID,"-",1:n())) %>% 
  st_transform(crs = "EPSG:4326")

# set crs
crs <- crs(patches)
```

## Calculate Landscape Patterns

### Function to Track Patch Changes
```{r get_patches, message=FALSE, warning=FALSE}
get_patches <- function(fire_name){
  print(fire_name)
  
  # select last timepoint's raster
  rast_fire <- rast(str_c("data/prediction_rasters/",fire_name,"_rf_t9.tif"))

  # use the high severity patches as t0
  t0_patches <- patches %>% 
    filter(Fire_ID == fire_name) %>% 
    dplyr::select(Fire_ID, Patch_ID,ecoregn,ptch_fr) %>% 
    mutate(t0_area = as.numeric(st_area(.))/10000)
  
  # convert the t9 raster to polygons
  t9_patches <- as.polygons(rast_fire) %>% 
    st_as_sf() %>% 
    st_cast(to = "POLYGON") %>% 
    st_transform(crs = crs)
  
  # join the t9 polygons to the original t0 patches to track decomposition
  conversion_df <- st_join(t9_patches,t0_patches,left = TRUE,largest = TRUE) %>% 
    st_intersection(.,st_union(t0_patches))%>% 
    mutate(t9_area = as.numeric(st_area(.))/10000) %>% 
    st_drop_geometry()
    
    # export csv
  write_csv(conversion_df,str_c("data/patch_fate/",fire_name,".csv"))
}
```

### Apply Function to All Fires

```{r map get_patches, message=FALSE, warning=FALSE,eval = FALSE}
# get the decomposition df for each fire 
map(fire_list,get_patches)
```

### Create Dataset for Each Forest Type

```{r get patch dfs, message=FALSE, warning=FALSE,cache = TRUE}
# import and join all decomposition csvs
patch_data <- do.call(bind_rows,lapply(list.files(path = "data/patch_fate/t0t9", pattern = "_t0t9", all.files=TRUE, full.names=TRUE),read_csv))
```


```{r get patch dfs by ftype, message=FALSE, warning=FALSE,cache = TRUE}
# filter data to each forest type, and prepare data for alluvial plot
patches_df <- patch_data %>% 
  filter(ptch_fr == "Douglas-Fir") %>% 
  make_long(t0_group,t9_group,value = t9_area)
patches_df$node <- factor(patches_df$node,levels = c("forested","0-50ha","50-100ha","100-500ha","500-1000ha","1000ha+"))
patches_df$next_node <- factor(patches_df$next_node,levels = c("forested","0-50ha","50-100ha","100-500ha","500-1000ha","1000ha+"))
patches_df$value <- as.numeric(patches_df$value)

patches_lp <- patch_data %>% 
  filter(ptch_fr == "Lodegepole Pine")%>% 
  make_long(t0_group,t9_group,value = t9_area)
patches_lp$node <- factor(patches_lp$node,levels = c("forested","0-50ha","50-100ha","100-500ha","500-1000ha","1000ha+"))
patches_lp$next_node <- factor(patches_lp$next_node,levels = c("forested","0-50ha","50-100ha","100-500ha","500-1000ha","1000ha+"))
patches_lp$value <- as.numeric(patches_lp$value)

patches_fs <- patch_data %>% 
  filter(ptch_fr == "Fir-Spruce")%>% 
  make_long(t0_group,t9_group,value = t9_area)
patches_fs$node <- factor(patches_fs$node,levels = c("forested","0-50ha","50-100ha","100-500ha","500-1000ha","1000ha+"))
patches_fs$next_node <- factor(patches_fs$next_node,levels = c("forested","0-50ha","50-100ha","100-500ha","500-1000ha","1000ha+"))
patches_fs$value <- as.numeric(patches_fs$value)
```

## Assess Landscape-Level Recovery

### By Forest Type

```{r reforested ftype, message=FALSE,warning=FALSE}
# percent of total area reforested by patch forest type

patch_data$t0_group <- factor(patch_data$t0_group,levels = c("0-50ha","50-100ha","100-500ha","500-1000ha","1000ha+"))

patch_data_type <- patch_data %>% 
  filter(ptch_fr %in% c("Fir-Spruce","Douglas-Fir","Lodegepole Pine")) %>% 
  dplyr::select(1:7) %>% 
  group_by(Patch_ID,t9_status,ptch_fr) %>% 
  summarize(t0_area = mean(t0_area),
            t9_area = sum(t9_area)) %>% 
  mutate(t9_status = case_when(t9_status == 1 ~ "unforested",
                               t9_status == 2 ~ "forested")) %>% 
  filter(t9_status == "forested")%>% 
  ungroup() %>% 
  group_by(ptch_fr) %>% 
  summarize(reforested_area_ha = sum(t9_area),
            burnt_area_ha = sum(t0_area),
            percent_reforested = round(reforested_area_ha/burnt_area_ha,2)) %>% 
  rename(forest_type = ptch_fr)


# table of proportion reforested by forest type
nice_table(patch_data_type, separate.header = TRUE,width =1)
```

### By Patch Size

```{r reforested size class, message=FALSE,warning=FALSE}
# percent of total area reforested by patch size class
patch_data_size <- patch_data %>% 
  filter(ptch_fr %in% c("Fir-Spruce","Douglas-Fir","Lodegepole Pine")) %>% 
  dplyr::select(1:7) %>% 
  group_by(Patch_ID,t9_status,t0_group) %>% 
  summarize(t0_area = mean(t0_area),
            t9_area = sum(t9_area)) %>% 
  mutate(t9_status = case_when(t9_status == 1 ~ "unforested",
                               t9_status == 2 ~ "forested")) %>% 
  filter(t9_status == "forested")%>% 
  ungroup() %>% 
  group_by(t0_group) %>% 
  summarize(reforested_area_ha = sum(t9_area),
            burnt_area_ha = sum(t0_area),
            percent_reforested = round(reforested_area_ha/burnt_area_ha,2)) %>% 
  rename(patch_size_class = t0_group)

# table of proportion reforested by size class
nice_table(patch_data_size, separate.header = TRUE,width =1)
```

## Calculate Fragmentation

```{r fragmentation function, message=FALSE,warning=FALSE}
## function to identify the proportion of area that has moved to a smaller size class
get_fragmentation <- function(long_df,ftype){
  frag <- long_df %>% 
    drop_na() %>% 
    filter(next_node != "forested")  %>% 
    group_by(node,next_node) %>% 
    summarize(class_area = sum(value)) %>% 
    mutate(total_area = sum(class_area)) %>% 
    ungroup() %>% 
    mutate(percent_frag = class_area/total_area,
           ptch_fr = ftype) 
}
```

```{r fragmentation map, message=FALSE,warning=FALSE}
# identify fragmentation separately across all three forest types
frag_df <- rbind(get_fragmentation(patches_lp,"Lodgepole Pine"),
                 get_fragmentation(patches_fs,"Fir-Spruce"),
                 get_fragmentation(patches_df,"Douglas-fir")) %>% 
  filter(node !="0-50ha")
```

```{r calc frag perc, message=FALSE,warning=FALSE}
# clean up fragmentation dataset
no_frag <- frag_df %>% 
  mutate(transition = node != next_node) %>% 
  group_by( ptch_fr,transition) %>% 
  summarize(transitioned_area = sum(class_area)) %>% 
  mutate(total_area = sum(transitioned_area),
         percent_transition = round(transitioned_area/total_area,2)) %>% 
  filter(transition == TRUE) %>% 
  rename(forest_type = ptch_fr) %>% 
  dplyr:: select(-transition)

# table of the proportion of area transitioning to smaller size classes for each forest type
nice_table(no_frag, separate.header = TRUE,width =1)
```

## Plot

### Alluvial Plot

```{r alluvial, message=FALSE, warning=FALSE}
# prepare alluvial plots for each forest type

# order by the node category
patches_lp <- patches_lp[order(patches_lp$next_node),]

# used to block the forested patch flows from color set
alphas2 <- c(rep(0,2800),rep(0.8,8600))

# used to block unforested patch flows from the null color set
colors <- c(rep("gray",2800),rep("#fc8d59",8600))
alphas <- c(rep(.4,2800),rep(0,8600))
               
plot_lp<- ggplot(patches_lp, aes(x = x, next_x = next_x, node = node, next_node = next_node, fill = factor(node), label = node,value = value)) +
  geom_alluvial(flow.alpha = alphas2,width =.28)+# plots colored flows for the unforested patches
  geom_alluvial(flow.alpha = alphas,flow.fill=colors,width =.28) +# plots null flows for the forested patches
  geom_alluvial_text(size = 4, color = "black") +# adds node label
  scale_fill_manual(values = c('#fef0d9','#fc8d59','#b30000','#fdcc8a','#e34a33',"gray")) + # adds base node colors
  theme_alluvial(base_size = 14) +# sets font/plot size
  labs(x = NULL, y= NULL,title = "Lodgepole Pine") +# labels
  theme(legend.position = "none")+# remove legend title
  theme(axis.text.x = element_text(vjust = 6)) + # centers the plot title, adjust x axis labels
  scale_x_discrete(labels = c('0 Years\nPost-Fire','30 Years\nPost-Fire'),expand = c(0, 0))  +# relabels the x axis names
  theme(panel.grid = element_blank(),panel.border = element_blank()) # remove excess white space

patches_df <- patches_df[order(patches_df$next_node),]

alphas2 <- c(rep(0,2800),rep(0.8,6200))
colors <- c(rep("gray",2800),rep("#fc8d59",6200))
alphas <- c(rep(.4,2800),rep(0,6200))
               
plot_df<- ggplot(patches_df, aes(x = x, next_x = next_x, node = node, next_node = next_node, fill = factor(node), label = node,value = value)) +
  geom_alluvial(flow.alpha = alphas2,width =.28)+
  geom_alluvial(flow.alpha = alphas,flow.fill=colors,width =.28) +
  geom_alluvial_text(size = 4, color = "black") +
  scale_fill_manual(values = c('#fef0d9','#fc8d59','#b30000','#fdcc8a','#e34a33',"gray")) + 
  theme_alluvial(base_size = 14) +
  labs(x = NULL, y= NULL,title = "Douglas-fir") +
  theme(legend.position = "none") +
  theme(plot.subtitle = element_text(hjust = 0.5), axis.text.x = element_text(vjust = 6)) +
  scale_x_discrete(labels = c('0 Years\nPost-Fire','30 Years\nPost-Fire'), expand = c(0, 0))  +
  theme(panel.grid = element_blank(),
        panel.border = element_blank())

patches_fs <- patches_fs[order(patches_fs$next_node),]

colors <- c(rep("gray",2800),rep("#fc8d59",9200))
alphas <- c(rep(.4,2800),rep(0,9200))
alphas2 <- c(rep(0,2800),rep(0.8,9200))
               
plot_fs <-ggplot(patches_fs, aes(x = x, next_x = next_x, node = node, next_node = next_node, fill = factor(node), label = node,value = value)) +
  geom_alluvial(flow.alpha = alphas2,width =.28)+
  geom_alluvial(flow.alpha = alphas,flow.fill=colors,width =.28) +
  geom_alluvial_text(size = 4, color = "black") +
  scale_fill_manual(values = c('#fef0d9','#fc8d59','#b30000','#fdcc8a','#e34a33',"gray")) + 
  theme_alluvial(base_size = 14) +
  labs(x = NULL, y= NULL,title = "Fir-Spruce") +
  theme(legend.position = "none") +
  theme(plot.subtitle = element_text(hjust = 0.5), axis.text.x = element_text(vjust = 6)) +
  scale_x_discrete(labels = c('0 Years\nPost-Fire','30 Years\nPost-Fire'), expand = c(0, 0))  +
  theme(panel.grid = element_blank(),
        panel.border = element_blank())
```

### All Three Forest Types

```{r, fig.width = 14,fig.height = 6}
# plot the three plots together
grid.arrange(plot_df,plot_lp,plot_fs,ncol=3, left = "Total Patch Area (ha)")
```
