==> 14_patchmetrics.Rmd <==

# Recovery Likelihood

## Set Up

### Libraries
```{r libraries12, message=FALSE, warning=FALSE}
library(tidyverse)
library(terra)
library(sf)
library(gridExtra)
library(car)
library(exactextractr)
library(elevatr)
library(rms)
library(rempsyc)
library(kableExtra)

sf_use_s2(FALSE)
```

### Import High-Severity Patches

```{r import patch6, message=FALSE, warning=FALSE,results='hide'}
# fire list
fire_list <- c("Fire_7_1988","Fire_9_1988","Fire_10_1988","Fire_11_1988","Fire_12_1988","Fire_13_1988","Fire_14_1988","Fire_15_1988","Fire_16_1988","Fire_18_1988","Fire_19_1988","Fire_20_1988","Fire_22_1988","Fire_23_1988","Fire_25_1988","Fire_26_1988","Fire_28_1988","Fire_29_1988","Fire_31_1988","Fire_32_1989","Fire_33_1989","Fire_35_1989","Fire_38_1989","Fire_41_1989","Fire_42_1989","Fire_48_1990","Fire_49_1991","Fire_50_1991","Fire_51_1991","Fire_54_1991","Fire_1_1988","Fire_2_1988","Fire_3_1988","Fire_4_1988")

# import high-severity patches
patches <- st_read("data/patches/highsev_patches.shp") %>%
  mutate(Patch_ID = str_c(Fire_ID,"-",1:n())) %>%
  st_transform(crs = "EPSG:4326")

# set crs
crs <- crs(patches)
```

## Prepare Patch Data

### Calculate Percent Reforested

```{r prop forested, message=FALSE, warning=FALSE}
# import and join patch fragmentation data
patch_fragmentation <- do.call(bind_rows,lapply(list.files(path = "data/patch_fate/t0t9", pattern = "t0t9", all.files=TRUE, full.names=TRUE),read_csv))

# calculate the proportion of area recolonized by conifers after 30 years
patch_recovery <- patch_fragmentation %>%
  group_by(Patch_ID,t9_status) %>%
  summarise(t0_area = mean(t0_area),
            t9_area = round(sum(t9_area),5))%>%
  pivot_wider(values_from = t9_area,names_from=t9_status) %>%
  rename(t9_area_unforested = 4,
         t9_area_forested = 3)%>%
  replace(is.na(.), 0) %>%
  mutate(perc_forested = t9_area_forested/(t0_area))
```

### Calculate Patch Area and Edge

```{r area and edge , message=FALSE, warning=FALSE}
patch_metrics <- patches %>%
  dplyr::select(Fire_ID,Patch_ID,ptch_fr,ecoregn) %>%
  st_transform(3488) %>%
  mutate(patch_area = round(as.numeric(st_area(.))/10000,4),
         perim_m = as.numeric(round(st_length(st_cast(geometry,to = "MULTILINESTRING")),4)),
         area_perim_rat = patch_area/perim_m) %>%
  st_buffer(-100) %>%
  mutate(core_area = round(as.numeric(st_area(.))/10000,4),
         perc_core = round(core_area / patch_area,5))%>%
  st_drop_geometry()
```

### Calculate Patch Climate

```{r terraclim, message=FALSE, warning=FALSE,cache=TRUE,results='hide'}
# import terraclim climate data, 3 year means before fire event
# for each patch, extract the mean value for each climate variable
get_climate <- function(var){
  extracted_var <- list.files(path = "data/terraclim/", pattern = var, all.files=TRUE, full.names=TRUE) %>%
    rast() %>%
    mean() %>% 
    extract(.,patches %>% st_transform(crs= crs(.)),fun = mean,weights = TRUE, exact = TRUE) %>%
    dplyr::select(mean)
  
  colnames(extracted_var) <- var
  
  return(extracted_var)
}

# list of climate variables
climate_variables <- c("def","pdsi","swe","ppt","pet","tmax","soil","vpd")

# extract mean climate values for each patch and assign to patch ID
climate_df <- map(climate_variables,get_climate)

patch_climate <- do.call(cbind,climate_df) %>% 
  cbind(Patch_ID = patches$Patch_ID,.)
```

### Calculate Latitude

```{r calc latitude, message=FALSE, warning=FALSE}
patch_latitude <- patches %>% 
  mutate(latitude = st_coordinates(st_centroid(geometry))[,2]) %>% 
  dplyr:: select(Patch_ID, latitude) %>% 
  st_drop_geometry()
```

### Calculate Topography

```{r calc topo, message=FALSE, warning=FALSE,cache=TRUE}
extract_topo <- function(fire_name){
  
  print(paste0("Starting Fire ",fire_name))
  
  # filter the patches for this fire
  fire_patches <- patches %>% 
    filter(Fire_ID==fire_name) %>% 
    st_transform(crs=4326)

  dem <- get_elev_raster(fire_patches,z=11)
  aspect <- terrain(dem, opt = "aspect",unit = "degrees")
  slope <- terrain(dem,opt='slope', unit='degrees')
  ccaspect <- cos(aspect)
  
  # get the mean topographic values for each patch in this fire
  extracted_elev <- left_join(fire_patches, exact_extract(dem,fire_patches, append_cols = TRUE, max_cells_in_memory = 3e+08, fun = "mean")) %>% 
    st_drop_geometry() %>% 
    rename(elevation = mean) %>% 
    dplyr::select(Patch_ID, elevation)
  extracted_aspect <- left_join(fire_patches, exact_extract(ccaspect,fire_patches, append_cols = TRUE, max_cells_in_memory = 3e+08, fun = "mean")) %>% 
    st_drop_geometry() %>% 
    rename(aspect = mean) %>% 
    dplyr::select(aspect)
  extracted_slope <- left_join(fire_patches, exact_extract(slope,fire_patches, append_cols = TRUE, max_cells_in_memory = 3e+08, fun = "mean")) %>% 
    st_drop_geometry() %>% 
    rename(slope = mean) %>% 
    dplyr::select(slope)
  
  extracted_data <- cbind(extracted_elev,extracted_aspect,extracted_slope)

  # export
  return(extracted_data)
}
```

```{r map topo, message=FALSE, results='hide'}
# map extraction function across all fires
extracted_fires <- map(fire_list,extract_topo)

patch_topography <- do.call(rbind,extracted_fires) 
```

### Join Patch Metrics and Recovery Data

```{r patch metrics joining, message=FALSE, warning=FALSE}
# combine all predictor datasets, set recovery threshold to 80% forested
log_recovery <- list(patch_metrics,patch_recovery,patch_latitude,patch_climate,patch_topography) %>% 
  reduce(full_join, by='Patch_ID') %>% 
  drop_na() %>% 
  filter(ptch_fr %in% c("Douglas-Fir","Lodegepole Pine","Fir-Spruce")) %>% 
  mutate(log_recovered = case_when(perc_forested >= 0.8 ~ 1,
                                   perc_forested < 0.8 ~ 0),
         area_perim_rat = area_perim_rat*10000,
         ecoregn = as.factor(ecoregn),
         ptch_fr= as.factor(ptch_fr)) %>% 
  st_drop_geometry()
```

## Model Recovery Likelihood

### Prepare Model Data

```{r model data prep, message=FALSE, warning=FALSE}
# data prep required for lrm model
dd <- datadist(log_recovery)
options(datadist="dd")
```

### Create Logistic Model

```{r lrm model, message=FALSE, warning=FALSE}
# logistic regression model with lrm
log_model <- lrm(log_recovered ~ ptch_fr + area_perim_rat + patch_area + elevation + slope + aspect + def + swe
                  ,data = log_recovery)
```

```{r lrm object, message=FALSE, warning=FALSE}
# lrm model object
log_model
```
```{r model summaries, message=FALSE, warning=FALSE}
#model summary odds ratio
kable(summary(log_model,def = c(1,2),swe = c(1,2),area_perim_rat = c(1,2),patch_area = c(1,2),elevation = c(1,2),slope = c(1,2),aspect = c(1,2)))

# model summary IQR odds ratio
kable(summary(log_model))
```

### Odds Ratio Table

```{r odds table, message=FALSE, warning=FALSE, eval = FALSE}
# bring in odds ratio table
odds_table <- read_csv("data/recovery/oddsratios_lrm.csv") %>% 
  dplyr::select(1:8)

# function to show data to correct number of digits
fun <- function(x) {formatC(x, format = "f", digits = 3)}

# create table 
odds <- nice_table(odds_table, separate.header = TRUE,width = 1,col.format.custom = c(3,5),format.custom = "fun")

# export to word 
# flextable::save_as_docx(odds, path = "data/recovery/odds_table.docx")
```
